+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/bounties/[bountyId]/bounty-submission-status-badges.ts
@@ -1,0 +1,14 @@
+export const BOUNTY_SUBMISSION_STATUS_BADGES = {
+  pending: {
+    label: "Pending",
+    variant: "pending",
+  },
+  approved: {
+    label: "Approved",
+    variant: "success",
+  },
+  rejected: {
+    label: "Rejected",
+    variant: "error",
+  },
+} as const;
